+++ app/src/main/java/com/chs/your_body_profile/domain/model/FoodDetailInfo.kt
@@ -2,7 +2,7 @@ package com.chs.your_body_profile.domain.model
 
 data class FoodDetailInfo(
     val name: String,
-    val servingWeight: Int = 0,
+    val servingWeight: Float = 0f,
     val calorie: Float = 0f,
     val carbohydrate: Float = 0f,
     val fat: Float = 0f,
